+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Change Log
 =====
 * Added PublicKey Key Types for differentiating between keys that can encrypt, sign or perform both actions.
 * Added the capability for separate Encryption and Signature keys for entities as an enabler for FIPS 140-3
+* Added totp functionality
+* Updated CLI with generateUserTotp, removeUserTotp, verifyUserTotp
 
 4.0.0
 =====